% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam.R
\name{adam_update}
\alias{adam_update}
\title{Perform an iteration of Adam otpimization}
\usage{
adam_update(g, adam_state)
}
\arguments{
\item{g}{The gradient for the current minibatch}

\item{adam_state}{Created by init_adam}
}
\description{
Perform an iteration of Adam otpimization
}
