\name{AdaptiveLearningRate-package}
\alias{AdaptiveLearningRate-package}
\alias{AdaptiveLearningRate}
\docType{package}
\title{\packageTitle{AdaptiveLearningRate}}
\description{\packageDescription{AdaptiveLearningRate}}
\section{Package Content}{\packageIndices{AdaptiveLearningRate}}
\author{\packageAuthor{AdaptiveLearningRate}}
\section{Maintainer}{\packageMaintainer{AdaptiveLearningRate}}
\keyword{package}
