% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_glm_sgd.R
\name{sgd_glm}
\alias{sgd_glm}
\title{Fit negative binomial (or Poisson) regression using Adam}
\usage{
sgd_glm(
  X,
  y,
  w = numeric(nrow(X)) + 1,
  o = numeric(nrow(X)),
  b = numeric(ncol(X)),
  conc = 10,
  learn_beta = T,
  learn_conc = T,
  batch_size = 10000,
  epochs = 1000,
  loglik_tol = 0.1,
  verbosity = 0,
  ...
)
}
\arguments{
\item{X}{NxP matrix of samples x covariates}

\item{y}{N-vector of observed counts}

\item{w}{Optional N-vector of observation weights}

\item{b}{Optional P-vector of coefficients. This is the initialization if b is being learned,}

\item{conc}{Optional concentration parameter. This is the initialization if conc is being learned.}

\item{learn_beta}{Whether to optimize the coefficients.}

\item{learn_conc}{Whether to optimize the concentration parameter.}

\item{batch_size}{For Adam}

\item{epochs}{Number of sweeps through dataset}

\item{loglik_tol}{Once the loglikelihood changes by less than this, stop.}

\item{verbosity}{Integer between 0 (only print warnings) and 3 (print at every iteration of optimization)}

\item{...}{passed to init_adam so can be used to specify e.g. learning_rate.}
}
\value{
Vector of log likelihoods through optimization (one per epoch).

adam_state list.
}
\description{
Fit negative binomial (or Poisson) regression using Adam
}
