% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_glm_sgd.R
\name{smart_fit_nb_glm}
\alias{smart_fit_nb_glm}
\title{Fit negative binomial (or Poisson) regression using Adam}
\usage{
smart_fit_nb_glm(X, y, w, o, consider_poisson = T, verbosity = 0, ...)
}
\arguments{
\item{X}{NxP matrix of samples x covariates}

\item{y}{N-vector of observed counts}

\item{w}{Optional N-vector of observation weights}

\item{o}{Offsets}

\item{consider_poisson}{Whether to compare the final NB GLM fit to the Poisson GLM and choose the later if it has higher loglikelihood. .}

\item{verbosity}{Integer between 0 (only print warnings) and 3 (print at every iteration of optimization)}

\item{...}{passed to sgd_glm}
}
\description{
Fit negative binomial (or Poisson) regression using Adam
}
