% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_glm_sgd.R
\name{adaglm}
\alias{adaglm}
\title{Main function for fitting negative binomial or Poisson GLM using Adam.}
\usage{
adaglm(
  formula,
  data,
  weights,
  conc = NULL,
  return_model = FALSE,
  return_x = FALSE,
  return_y = FALSE,
  contrasts = NULL,
  verbosity = 0,
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which adaglm is called.}

\item{weights}{an optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector.}

\item{conc}{Concentration parameter. NULL to learn it, or can be fixed to some value. Setting conc=Inf gives Poisson GLM.}

\item{return_model}{Whether to return the model frame.}

\item{return_x}{Whether to return the model matrix}

\item{return_y}{Whether to return the response vector}

\item{contrasts}{an optional list. See the contrasts.arg of model.matrix.default.}

\item{verbosity}{Integer between 0 (only print warnings) and 3 (print at every iteration of optimization)}

\item{...}{Passed to smart_fit_nb_glm, then sgd_glm, and finally init_adam. Can control the optimization through `batch_size`, `epochs`, `loglik_tol`, `beta1`, `beta2`, `epsilon` and `learning_rate`.}
}
\description{
This aims to have roughly the same functionality as `base::glm`. Offsets are not currently supported however, and currently only log link can be used.
}
